Remove no longer needed AboutTapToPayView and replace with TapToPayEducation
