+++ WooCommerce/Classes/ViewRelated/Dashboard/Settings/CardReadersV2/SetUpTapToPayCompleteViewController.swift
@@ -92,7 +92,7 @@ struct SetUpTapToPayCompleteView: View {
         }
         .padding()
         .sheet(isPresented: $showingAboutTapToPay) {
-            AboutTapToPayViewInNavigationView(viewModel: AboutTapToPayViewModel(shouldAlwaysHideSetUpTapToPayButton: true))
+            TapToPayEducationView(viewModel: .init(flow: .onboarding, completion: { _ in }))
         }
     }
 }
