+++ WooCommerce/Classes/ViewRelated/Dashboard/Settings/CardReadersV2/Tap to Pay Education/TapToPayEducationStepViewModel.swift
@@ -4,12 +4,12 @@ final class TapToPayEducationStepViewModel {
     let title: String
     let imageName: String
     let descriptionSteps: [String]
-    let limit: AboutTapToPayContactlessLimitViewModel?
+    let limit: TapToPayEducationContactlessLimitViewModel?
 
     init(title: String,
          imageName: String,
          descriptionSteps: [String],
-         limit: AboutTapToPayContactlessLimitViewModel? = nil) {
+         limit: TapToPayEducationContactlessLimitViewModel? = nil) {
         self.title = title
         self.imageName = imageName
         self.descriptionSteps = descriptionSteps
@@ -19,7 +19,7 @@ final class TapToPayEducationStepViewModel {
     init(title: String,
          imageName: String,
          description: String,
-         limit: AboutTapToPayContactlessLimitViewModel? = nil) {
+         limit: TapToPayEducationContactlessLimitViewModel? = nil) {
         self.title = title
         self.imageName = imageName
         self.descriptionSteps = [description]
