+++ WooCommerce/Classes/ViewRelated/Dashboard/Settings/In-Person Payments/Payments Menu/InPersonPaymentsMenu.swift
@@ -62,40 +62,27 @@ struct InPersonPaymentsMenu: View {
                                 onboardingUseCase: viewModel.onboardingUseCase)
                         }
 
-                        if viewModel.isTapToPayEducationEnabled {
-                            PaymentsRow(image: Image(uiImage: .infoOutlineImage),
-                                        title: Localization.aboutTapToPayOnIPhone)
-                            .accessibilityAddTraits(.isButton)
-                            .onTapGesture {
-                                viewModel.aboutTapToPayTapped()
+                        PaymentsRow(image: Image(uiImage: .infoOutlineImage),
+                                    title: Localization.aboutTapToPayOnIPhone)
+                        .accessibilityAddTraits(.isButton)
+                        .onTapGesture {
+                            viewModel.aboutTapToPayTapped()
+                        }
+                        .sheet(isPresented: $viewModel.presentAboutTapToPay,
+                               onDismiss: {
+                            Task { @MainActor in
+                                await viewModel.onAppear()
                             }
-                            .sheet(isPresented: $viewModel.presentAboutTapToPay,
-                                   onDismiss: {
-                                Task { @MainActor in
-                                    await viewModel.onAppear()
+                        }) {
+                            TapToPayEducationView(viewModel: .init(flow: .about,
+                                                                   completion: { result in
+                                switch result {
+                                case .setUpTapToPay:
+                                    viewModel.presentSetUpTryOutTapToPay = true
+                                default:
+                                    break
                                 }
-                            }) {
-                                TapToPayEducationView(viewModel: .init(flow: .about,
-                                                                       completion: { result in
-                                    switch result {
-                                    case .setUpTapToPay:
-                                        viewModel.presentSetUpTryOutTapToPay = true
-                                    default:
-                                        break
-                                    }
-                                }))
-                            }
-                        } else {
-                            Button {
-                                viewModel.aboutTapToPayTapped()
-                            } label: {
-                                PaymentsRow(image: Image(uiImage: .infoOutlineImage),
-                                            title: Localization.aboutTapToPayOnIPhone,
-                                            isActive: $viewModel.presentAboutTapToPay) {
-                                    AboutTapToPayView(viewModel: viewModel.aboutTapToPayViewModel)
-                                }
-                            }
-                            .buttonStyle(.scrollViewRow)
+                            }))
                         }
                     } header: {
                         Text(Localization.tapToPaySectionTitle.uppercased())
